Move and cleanup ExpansionCheckbox
